Move component to different file, make date black
